// frontend/src/lib/generated/types.gen.ts
export type ElementsLabelEntry = {
  /**
   * A string field
   */
  icon?: string;
  iconPosition: IconPositionEnum;
  /**
   * A text field
   */
  text: string;
  /**
   * A string field
   */
  ariaDescription: string;
  /**
   * Component instance ID
   */
  id?: number;
};

export type ElementsHeaderEntry = {
  alignment: AlignmentEnum;
  /**
   * A boolean field
   */
  promoteHeaderIcon: boolean | null;
  /**
   * A component field
   */
  header?: ElementsLabelEntry;
  /**
   * A component field
   */
  subheader?: ElementsLabelEntry;
  /**
   * Component instance ID
   */
  id?: number;
};

export type ElementsButtonEntry = {
  /**
   * A string field
   */
  url: string;
  /**
   * A boolean field
   */
  openInNewTab: boolean | null;
  /**
   * A component field
   */
  label?: ElementsLabelEntry;
  /**
   * Component instance ID
   */
  id?: number;
};
// ...
export enum DirectionEnum {
  LTR = "ltr",
  RTL = "rtl",
}

/**
 * An enum field
 */
export enum IconPositionEnum {
  BEFORE_TEXT = "before_text",
  AFTER_TEXT = "after_text",
}

/**
 * An enum field
 */
export enum AlignmentEnum {
  CENTER = "center",
  LANGUAGE_DIRECTION = "language-direction",
}